Use an exact match for backwards-compatible-json.
Previously the doc said it was a prefix match, but that is both
error-prone and slightly complex. It is safer and easier to use an exact
string match. We use a magic URI so that it may optionally serve
documentation for this format.

Also add a warning that backwards compatibility should be disabled by
default.