Replace magic payloadType with explicit sigType.
Previously the switch between normal and "backwards compatible json"
mode was determined by a magic `payloadType`. The downside of this
approach was that it was undiscoverable: readers would not know this
magic value existed.

Now this behavior is made explicit via a `signatures.sigType` field. The
new string better explains the behavior, which is that the payload must
be JSON and that `payloadType` is unauthenticated.