Add "design requirements" section.
This helps readers understand what problems we are trying to solve and
why we chose these particular decisions. It also helps us assess
potential future changes.

NOTE: The "multi-message file format" is not yet in the spec. I am
adding it here because we do need to eventually define that.