v1.0.2: explicitly forbid reparsing the envelope
It is a security mistake for an implementation to parse the envelope,
verify the signature, and then re-parse the envelope to pull out the
payload, since the two different parsings could result in different
payloads.

This was already implicitly disallowed in protocol.md according to the
"to verify" instructions, but a reader who is not careful might miss
this important bit.

Now we call it out explicitly on both protocol.md and envelope.md, to
avoid any ambiguity.

Signed-off-by: Mark Lodato <lodato@google.com>